+++ include/Move_command.hpp
@@ -10,7 +10,7 @@
 
 #include <Manifold.hpp>
 
-template <std::int_fast64_t dimension>
+template <std::size_t dimension>
 class Move_command;
 
 template <>
@@ -27,7 +27,7 @@ class Move_command<3>
   };
 
   using Move_queue   = std::vector<Move_type>;
-  using Move_tracker = std::array<std::int_fast64_t, 5>;
+  using Move_tracker = std::array<std::int_fast32_t, 5>;
 
   /// @brief Default constructor
   Move_command()
